[EI-133] [web] chore: Update metadata-dialog-menu import paths
