+++ web/__tests__/components/numerical-guidance/metadata-dialog-menu.test.tsx
@@ -3,7 +3,7 @@ import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db.mock';
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import MetdataList from '@/app/ui/components/numerical-guidance/molecule/metadata-list';
-import MetadataDialogMenu from '@/app/ui/components/numerical-guidance/molecule/metadata-dialog-menu';
+import MetadataDialogMenu from '@/app/ui/components/numerical-guidance/molecule/metadata-dialog-menu/metadata-dialog-menu';
 import userEvent from '@testing-library/user-event';
 
 describe('MetadataDialogMenu', () => {
